requirements.txt
# integration and db
PyMuPDF
weaviate-client
transformers==4.46.2
sentence-transformers==3.4.1
nltk
tiktoken

# llm
litellm

# api
fastapi
uvicorn

# global
python-dotenv
